requirements.txt
requests==2.31.0
PyGithub==1.59.1
gitpython==3.1.40
